backend/requirements.txt
Flask==2.1.2
bcrypt==3.2.0
python-dotenv==0.19.0
PyMySQL==1.0.2
requests==2.26.0
Flask-JWT-Extended
